issue8: first cut redis deployment / integration
