+++ package.json
@@ -4,6 +4,7 @@
     "version": "0.1.0",
     "dependencies": {},
     "devDependencies": {
-        "serverless": "1.61.3"
+        "serverless": "1.61.3",
+        "serverless-python-requirements": "^5.1.0"
     }
 }
