--- requirements.txt
@@ -1,1 +1,0 @@
-boto3
