+++ src/requirements-dev.txt
@@ -1,0 +1,3 @@
+moto
+pytest
+boto3==1.14.23
