+++ src/requirements.txt
@@ -1,0 +1,1 @@
+redis==3.5.3
