--- tests/unit/handlers/test_rgbHandler.py
@@ -1,9 +1,0 @@
-import pytest
-import handlers.rgbHandler as h
-
-def test_getRgb():
-    '''Test rgb handler.'''
-    event = {'dummy': 'data'}
-    rgbResp = h.getRgb(event, {})
-    
-    assert rgbResp.get('statusCode') == 200
